Use milliseconds as default granularity for dateTime
