+++ generator/src/main/java/com/scottlogic/deg/generator/generation/StandardFieldValueSourceEvaluator.java
@@ -4,6 +4,7 @@ import com.scottlogic.deg.generator.FlatMappingSpliterator;
 import com.scottlogic.deg.generator.constraints.atomic.IsOfTypeConstraint;
 import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 import com.scottlogic.deg.generator.generation.fieldvaluesources.*;
+import com.scottlogic.deg.generator.generation.fieldvaluesources.DateTime.TemporalFieldValueSource;
 import com.scottlogic.deg.generator.restrictions.*;
 
 import java.math.BigDecimal;
